+++ packages/bundle/src/adaptiveCards/Attachment/SignInCardContent.tsx
@@ -1,11 +1,9 @@
-import { hooks } from 'botframework-webchat-component';
 import PropTypes from 'prop-types';
 import React, { FC } from 'react';
 import type { DirectLineSignInCard } from 'botframework-webchat-core';
 
 import CommonCard from './CommonCard';
-
-const { useStyleSet } = hooks;
+import useStyleSet from '../../hooks/useStyleSet';
 
 type SignInCardContentProps = {
   actionPerformedClassName?: string;