+++ packages/component/src/Styles/StyleSet/ModalDialog.ts
@@ -5,10 +5,12 @@ import {
   NOT_FORCED_COLORS_SELECTOR
 } from './Constants';
 
+import CSSTokens from '../CSSTokens';
+
 export default function createModalDialogStyleSet() {
   return {
     '&.webchat__modal-dialog': {
-      fontFamily: 'var(--webchat__font--primary)',
+      fontFamily: CSSTokens.FontPrimary,
       width: '100%',
 
       [NOT_FORCED_COLORS_SELECTOR]: {
@@ -18,11 +20,11 @@ export default function createModalDialogStyleSet() {
 
       '& .webchat__modal-dialog__box': {
         borderRadius: 2,
-        height: 'calc(100% - var(--webchat__padding--regular) * 2)',
+        height: `calc(100% - ${CSSTokens.PaddingRegular} * 2)`,
         overflow: 'hidden',
         margin: 'auto',
         maxWidth: '60%',
-        width: 'calc(100% - var(--webchat__padding--regular) * 2)',
+        width: `calc(100% - ${CSSTokens.PaddingRegular} * 2)`,
 
         [LIGHT_THEME_SELECTOR]: {
           // From Power BI:
@@ -49,7 +51,7 @@ export default function createModalDialogStyleSet() {
 
       '& .webchat__modal-dialog__close-button-layout': {
         float: 'right',
-        padding: 'var(--webchat__padding--regular)'
+        padding: CSSTokens.PaddingRegular
       },
 
       '& .webchat__modal-dialog__close-button': {
@@ -112,7 +114,7 @@ export default function createModalDialogStyleSet() {
       },
 
       '& .webchat__modal-dialog__body': {
-        margin: 'calc(var(--webchat__padding--regular) * 2)'
+        margin: `calc(${CSSTokens.PaddingRegular} * 2)`
       }
     }
   };