+++ packages/component/src/Styles/StyleSet/SendStatus.ts
@@ -1,10 +1,12 @@
+import CSSTokens from '../CSSTokens';
+
 export default function createSendStatusStyle() {
   return {
     '&.webchat__activity-status': {
-      color: 'var(--webchat__color--timestamp)',
-      fontFamily: 'var(--webchat__font--primary)',
-      fontSize: 'var(--webchat__font-size--small)',
-      marginTop: 'calc(var(--webchat__padding--regular) / 2)'
+      color: CSSTokens.ColorTimestamp,
+      fontFamily: CSSTokens.FontPrimary,
+      fontSize: CSSTokens.FontSizeSmall,
+      marginTop: `calc(${CSSTokens.PaddingRegular} / 2)`
     },
 
     '&.webchat__activity-status--slotted': {
@@ -16,7 +18,7 @@ export default function createSendStatusStyle() {
       alignItems: 'center',
 
       '&.webchat__activity-status__originator--has-link': {
-        color: 'var(--webchat__color--accent)'
+        color: CSSTokens.ColorAccent
       }
     }
   };