+++ packages/component/src/Styles/StyleSet/SlottedActivityStatus.ts
@@ -1,13 +1,15 @@
+import CSSTokens from '../CSSTokens';
+
 export default function createSlottedActivityStatus() {
   return {
     '&.webchat__slotted-activity-status': {
       alignItems: 'center',
       display: 'inline-flex',
       gap: 4,
-      marginTop: 'calc(var(--webchat__padding--regular) / 2)',
+      marginTop: `calc(${CSSTokens.PaddingRegular} / 2)`,
 
       '& .webchat__slotted-activity-status__pipe': {
-        fontSize: 'var(--webchat__font-size--small)'
+        fontSize: CSSTokens.FontSizeSmall
       }
     }
   };