+++ packages/component/src/hooks/useStyleSet.ts
@@ -1,5 +1,7 @@
+import type createStyleSet from '../Styles/createStyleSet';
 import useWebChatUIContext from './internal/useWebChatUIContext';
 
-export default function useStyleSet(): [any] {
-  return [useWebChatUIContext().styleSet];
+// TODO: Seems type of value is `string`.
+export default function useStyleSet(): readonly [Record<keyof ReturnType<typeof createStyleSet>, any>] {
+  return Object.freeze([useWebChatUIContext().styleSet] as const);
 }