+++ packages/component/src/types/internal/TypeOfArray.ts
@@ -1,7 +1,5 @@
 export type TypeOfArray<T extends unknown[] | Array<unknown> | ReadonlyArray<unknown>> = T extends (infer I)[]
   ? I
-  : T extends Array<infer I>
-  ? I
-  : T extends ReadonlyArray<infer I>
+  : T extends readonly (infer I)[]
   ? I
   : T;